backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
celery==5.3.6
redis==5.0.3
pydantic==2.6.3
python-dotenv==1.0.1
ffmpeg-python==0.2.0
websockets==12.0
aiofiles==23.2.1
